Fixed chapter change bug and added saxon build
The cli would not let you run excersized in any chapter other than 3.
This is the chapter I was reading when I decided to write the CLI.

Saxonpy will not run on a mac, so I am adding a custom build step to the
setup.py file that will be run with 'python setup install'.  This is not
done yet, since I do not have a mac to test it on.